# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  db.GetOrganisationTransfersTxResponse:
    properties:
      incoming_transfers:
        items:
          $ref: '#/definitions/db.IncomingTransfer'
        type: array
      outgoing_transfer:
        $ref: '#/definitions/db.OutgoingTransfer'
    type: object
  db.IncomingTransfer:
    properties:
      created_at:
        type: integer
      from_profile_name:
        type: string
      id:
        type: string
      organisation_description:
        type: string
      organisation_name:
        type: string
    type: object
  db.OutgoingTransfer:
    properties:
      created_at:
        type: integer
      id:
        type: string
      organisation_description:
        type: string
      organisation_name:
        type: string
      to_profile_name:
        type: string
    type: object
  organisation.createOrganisationInput:
    properties:
      description:
        example: The organisation for bugs!
        maxLength: 100
        type: string
      name:
        example: buggy org
        maxLength: 20
        type: string
    required:
    - description
    - name
    type: object
  organisation.organisationOutput:
    properties:
      created_at:
        example: 1710579130
        type: integer
      description:
        example: The organisation for bugs!
        type: string
      id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      name:
        example: buggy org
        type: string
      owner:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      updated_at:
        example: 1710579130
        type: integer
    type: object
  profile.changePasswordInput:
    properties:
      new_password:
        example: something1234
        maxLength: 32
        minLength: 8
        type: string
      old_password:
        example: something123
        type: string
    required:
    - new_password
    - old_password
    type: object
  profile.createProfileInput:
    properties:
      email:
        example: asrivatsa6@gmail.com
        maxLength: 32
        type: string
      name:
        example: abhinav
        maxLength: 32
        type: string
      password:
        example: something123
        maxLength: 32
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  profile.profileOutput:
    properties:
      created_at:
        example: 1710579130
        type: integer
      email:
        example: asrivatsa6@gmail.com
        type: string
      id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      name:
        example: abhinav
        type: string
      updated_at:
        example: 1710579130
        type: integer
      verified:
        example: false
        type: boolean
    type: object
...
  team.teamOutput:
    properties:
      created_at:
        example: 1710579130
        type: integer
      description:
        example: The Buggy testing team!
        type: string
      id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      name:
        example: testing team
        type: string
      organisation_description:
        example: The organisation for bugs!
        type: string
      organisation_name:
        example: buggy org
        type: string
      updated_at:
        example: 1710579130
        type: integer
    type: object
...
  transfer.transferInput:
    properties:
      email:
        example: asrivatsa6@gmail.com
        type: string
    required:
    - email
    type: object
  transfer.transferOutput:
    properties:
      completed:
        example: false
        type: boolean
      created_at:
        example: 1710579130
        type: integer
      from_profile:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      id:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      organisation:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      to_profile:
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
    type: object
  util.ErrorDetail:
    properties:
      location:
        type: string
      message:
        type: string
      value: {}
    type: object
  util.ErrorModel:
    properties:
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/util.ErrorDetail'
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: The comprehensive list of all Buggy APIs
  title: Buggy APIs
  version: "1.0"
paths:
  /organisation:
    get:
      description: Get an organisation, this API will return an error if the organisation
        does not exist.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organisation.organisationOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Get an organisation.
      tags:
      - organisation
    post:
      consumes:
      - application/json
      description: Create a new organisation, each profile may only have one organisation.
      parameters:
      - description: json
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/organisation.createOrganisationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organisation.organisationOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Create a new organisation.
      tags:
      - organisation
    put:
      consumes:
      - application/json
      description: Update an organisations name or description.
      parameters:
      - description: json
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/organisation.createOrganisationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organisation.organisationOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Update an organisations name or description.
      tags:
      - organisation
...
  /organisation/team/{team-id}:
    put:
      description: Update a team under a profile's organisation.
      parameters:
      - description: Team ID
        in: path
        name: team-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/team.teamOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Update a team under a profile's organisation.
      tags:
      - organisation
  /organisation/transfer:
    get:
      description: Get all incoming and outgoing organisation transfers.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.GetOrganisationTransfersTxResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Get all organisation transfers.
      tags:
      - organisation
    post:
      consumes:
      - application/json
      description: Create a new organisational transfer.
      parameters:
      - description: json
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/transfer.transferInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transfer.transferOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Create a new organisational transfer.
      tags:
      - organisation
  /organisation/transfer/{organisation-transfer-id}:
    delete:
      consumes:
      - application/json
      description: Delete an organisational transfer.
      parameters:
      - description: Organisation Transfer ID
        in: path
        name: organisation-transfer-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transfer.transferOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Delete an organisational transfer.
      tags:
      - organisation
...
  /profile:
    get:
      description: Get a profile from a token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Get a profile from a token.
      tags:
      - profile
    post:
      consumes:
      - application/json
      description: Create a new profile by giving a profile name, email and password.
        This API will generate a verification email which will contain a redirect
        link.
      parameters:
      - description: json
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/profile.createProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Create a new profile, emails must be unique.
      tags:
      - profile
...
  /profile/password/change:
    post:
      consumes:
      - application/json
      description: Change a profile password using the old password for confirmation.
      parameters:
      - description: json
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/profile.changePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorModel'
      summary: Change a profile password.
      tags:
      - profile
...
swagger: "2.0"